Create Covid Analysis Using Sql
This project analyzes global COVID-19 data using SQL, focusing on trends in infections, deaths, and vaccinations. It includes queries for death and infection rates, population impact, and vaccine rollout analysis using CTEs, temp tables, and views. Data is sourced from the CovidDeaths and CovidVaccination datasets. The goal is to uncover insights and prepare data for visualization in tools like Power BI or Tableau. This repository serves as a foundation for data-driven decision-making in public health.